refactor(file-picker): own the truncate-path pipe and drop dead surface
The publicly-exported truncate-path pipe imported PathSegment and the
path utils from the file-picker feature — an inverted dependency for a
pipe whose semantics (rootPath argument, breadcrumb segments) are
file-picker-specific. Move it under file-picker/; consumers import from
the package root, so the flattened export is unchanged.

Also drop code only tests were keeping alive: the never-called
updateSelectionFromItems(), and the popup's getFileInfo /
formatFileSize / getTypeDisplayName, which no template references —
the template already formats sizes through FileSizePipe.
