+++ projects/truenas-ui/src/lib/file-picker/index.ts
@@ -3,3 +3,4 @@ export * from './file-picker-popup.component';
 export * from './file-picker.interfaces';
 export * from './file-picker.utils';
 export * from './path-utils';
+export * from './truncate-path.pipe';
