+++ projects/truenas-ui/src/lib/file-picker/truncate-path.pipe.spec.ts
