+++ projects/truenas-ui/src/public-api.ts
@@ -94,7 +94,6 @@ export * from './lib/pipes/file-size/file-size.pipe';
 export * from './lib/pipes/label-markup/label-markup.pipe';
 export * from './lib/pipes/label-markup/label-text.pipe';
 export * from './lib/pipes/label-markup/label-markup.utils';
-export * from './lib/pipes/truncate-path/truncate-path.pipe';
 export * from './lib/spinner/spinner.component';
 export * from './lib/spinner/branded-spinner.component';
 export * from './lib/progress-bar/progress-bar.component';
